requirements.txt
Flask==2.3.3
flask-restx==1.1.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy>=3.0.0
flasgger==0.9.7.1
google-cloud-storage==2.10.0
gunicorn==20.1.0
requests==2.31.0 
psycopg2-binary==2.9.7
pgvector==0.1.8
opencv-python-headless==4.8.0.76
opencv-python==4.8.0.76
opencv-contrib-python==4.8.0.76
numpy==1.24.4
scipy==1.10.1
Pillow==9.5.0
python-dotenv
scikit-learn==1.2.2
fpdf==1.7.2
WeasyPrint==53.3
gunicorn
